+++ code/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [#995](https://github.com/InditexTech/weavejs/issues/995) Remove timeout to refresh token on Azure Web PubSub
+- [#996](https://github.com/InditexTech/weavejs/issues/996) Refresh reconnection attempts on Azure Web PubSub store once a connection is stablished
 
 ## [3.3.0] - 2026-03-18